requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
APIFlask==1.0.0
apispec==5.2.2
click==8.1.3
Flask==2.1.2
Flask-HTTPAuth==4.7.0
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
marshmallow==3.16.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.20.0
six==1.16.0
webargs==8.1.0
Werkzeug==2.1.2